// pos/src/app/models/m-country-master.ts
export class MCountryMaster {
    constructor() { }

    id?: number;
    countryCode?: string;
    countryName?: string;
    languageName?: string;
    decimalDigit?: number;
    decimalPlaces?: number;
    dateFormat?: string;
    dateSeparator?: string;
    negativeSign?: string;
    currencySeparator?: string;
    currency?: string;
    emailID?: string;
    creditLimitCheck?: boolean;
    allowMultipleTransaction?: boolean;
    allowPartialReceiving?: boolean;
    allowSaleAndRedemption?: boolean;
    orginCountry?: boolean;
    currencyID?: number;
    taxID?: number;
    nearByRoundOff?: number;
    currencyCode?: string;
    taxCode?: string;
    posTitle?: string;
    promotionRoundOff?: number;
    active?: boolean;

    //  List<ShiftMaster> Shiftlist 
}